requirements.txt
<<<<<<< HEAD
# Energy Recommendation System - Dependencies
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning and neural networks
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.6.0

# Time series and forecasting
statsmodels>=0.12.0
prophet>=1.0.0
sktime>=0.10.0

# Optimization libraries
cvxpy>=1.2.0
gurobipy>=9.5.0  # Optional: Commercial solver
ortools>=9.0.0   # Alternative: Open source solver

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Energy data processing
pandas-profiling>=3.2.0
great-expectations>=0.13.0

# Weather and climate data
xarray>=0.20.0
netcdf4>=1.5.0
pyproj>=3.2.0

# Testing and quality assurance
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0
pytest-benchmark>=3.4.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Documentation
jupyter>=1.0.0
jupyterlab>=3.0.0
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Configuration management
pyyaml>=5.4.0
python-dotenv>=0.19.0
hydra-core>=1.1.0

# Logging and monitoring
loguru>=0.5.0
mlflow>=1.20.0
wandb>=0.12.0  # Optional: Weights & Biases

# Data serialization and storage
joblib>=1.0.0
h5py>=3.1.0
zarr>=2.10.0

# Performance optimization
numba>=0.53.0
cython>=0.29.0
dask>=2021.9.0

# Energy system modeling
pandapower>=2.8.0  # Power system analysis
pypsa>=0.19.0      # Energy system optimization

# API and web services
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.0

# Development tools
pre-commit>=2.15.0
ipython>=7.25.0
ipykernel>=6.0.0
jupyter-contrib-nbextensions>=0.5.0

# Optional: GPU acceleration
# cupy>=9.0.0
# nvidia-ml-py3>=7.352.0

# Optional: Cloud deployment
# boto3>=1.20.0
# azure-storage-blob>=12.8.0
# google-cloud-storage>=2.0.0
=======
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization and Dashboard
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0

# Additional Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=1.0.0

# Web Framework and Server
flask>=2.3.0
gunicorn>=21.0.0

# Data Storage and Processing
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
>>>>>>> origin/visualization
